pyproject.toml
[project]
name = "image-generator"
version = "0.1.0"
description = "Generate images from prompts using Dreamlook API"
dependencies = ["requests", "openai", "python-dotenv>=1.0.1"]

[tool.ruff]
line-length = 88
